+++ gdnative-derive/src/variant.rs
